requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
openai>=1.3.0
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
requests>=2.31.0

